Updated and restructured callable_argument_helper.hpp to make it compile
- Added CMakeLists.txt
- Changed all include directives to relative paths to this repo
- Moved all source files to /include
- Restructured callable_argument_helper.hpp to fix all compiler errors
